requirements.txt
langgraph>=0.6.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
aerospike>=15.0.0
langgraph-checkpoint-aerospike>=0.1.0
langgraph-store-aerospike>=0.1.0
gremlinpython==3.7.2
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
